Refactored the query stuff. There was an issue with re-using queries for statements encountered when implementing the goSql library engine middleware. This breakdown of files is easier to interpret.
This also breaks down the interfaces and removes needless method calls and consolidates them down to, really two: getting the sqlquery, and interpolation of the query. The rest of the implementation is independent of these two components. This should facilitate implementations of any future Queryer/Parameterer interfaces in the future.